GPT-OSSC/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpt-oss-ws"
version = "0.1.0"
description = "Latent global workspace hooks and OpenAI-compatible serving stack for GPT-OSS-20B"
requires-python = ">=3.11"
authors = [{name = "OSSC Systems Team", email = "systems@ossc.local"}]
license = "Apache-2.0"
readme = "README.md"
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Operating System :: POSIX :: Linux",
  "Operating System :: Microsoft :: Windows",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence"
]

dependencies = [
  "accelerate>=0.33",
  "bitsandbytes>=0.43",
  "datasets>=2.20",
  "fastapi>=0.111",
  "faiss-cpu>=1.8.0",
  "huggingface-hub>=0.23",
  "numpy>=1.26",
  "pandas>=2.1",
  "matplotlib>=3.9",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "safetensors>=0.4",
  "requests>=2.32",
  "scikit-learn>=1.4",
  "sentence-transformers>=2.7",
  "sqlalchemy>=2.0",
  "starlette>=0.37",
  "tensorboard>=2.17",
  "torch",
  "transformers>=4.44",
  "tqdm>=4.66",
  "typer[all]>=0.12",
  "uvicorn>=0.30"
]

[project.optional-dependencies]
cli = ["rich>=13.7"]
test = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "pytest-timeout>=2.3"
]

[tool.setuptools]
package-dir = {"" = ""}
packages = [
  "gpt_oss_ws",
  "gpt_oss_ws.abliteration",
  "gpt_oss_ws.utils",
  "cli",
  "evals"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --timeout=120"
testpaths = ["tests"]

[tool.typer.plugins.auto-completion]
completion-install = "automatic"
